refactored nm folder utils
