+++ nemo/collections/nlp/metrics/squad_metrics.py
@@ -21,7 +21,7 @@ import collections
 from transformers.tokenization_bert import BasicTokenizer
 
 from nemo import logging
-from nemo.collections.nlp.utils.common_nlp_utils import get_tokens, normalize_answer
+from nemo.collections.nlp.utils.data_utils import get_tokens, normalize_answer
 
 __all__ = [
     'f1_score',
