+++ nemo/collections/nlp/nm/trainables/common/sequence_regression_nm.py
@@ -17,7 +17,7 @@
 from torch import nn as nn
 
 from nemo.backends.pytorch import MultiLayerPerceptron, TrainableNM
-from nemo.collections.nlp.nm.trainables.common.transformer.transformer_utils import transformer_weights_init
+from nemo.collections.nlp.utils.transformer_utils import transformer_weights_init
 from nemo.core import ChannelType, NeuralType, RegressionValuesType
 
 __all__ = ['SequenceRegression']
