+++ nemo/collections/nlp/nm/trainables/common/token_classification_nm.py
@@ -17,7 +17,8 @@
 from torch import nn as nn
 
 from nemo.backends.pytorch import MultiLayerPerceptron, TrainableNM
-from nemo.collections.nlp.nm.trainables.common.transformer.transformer_utils import gelu, transformer_weights_init
+from nemo.collections.nlp.utils.functional_utils import gelu
+from nemo.collections.nlp.utils.transformer_utils import transformer_weights_init
 from nemo.core import ChannelType, LogitsType, NeuralType
 
 __all__ = ['BertTokenClassifier', 'TokenClassifier']
