+++ nemo/collections/nlp/nm/trainables/common/transformer/transformer_encoders.py
@@ -24,7 +24,7 @@ from nemo.collections.nlp.nm.trainables.common.transformer.transformer_modules i
     PositionWiseFF,
     TwoStreamSelfAttention,
 )
-from nemo.collections.nlp.nm.trainables.common.transformer.transformer_utils import form_attention_mask
+from nemo.collections.nlp.utils.transformer_utils import form_attention_mask
 
 __all__ = []
