+++ nemo/collections/nlp/nm/trainables/common/transformer/transformer_generators.py
@@ -17,8 +17,8 @@
 import torch
 import torch.nn as nn
 
-from nemo.collections.nlp.nm.trainables.common.transformer.transformer_utils import NEG_INF
-from nemo.collections.nlp.utils.loss_utils import mask_padded_tokens
+from nemo.collections.nlp.utils.data_utils import mask_padded_tokens
+from nemo.collections.nlp.utils.transformer_utils import NEG_INF
 
 __all__ = []
