+++ nemo/collections/nlp/utils/data_utils.py
@@ -17,7 +17,7 @@
 import re
 import string
 
-__all__ = ['get_vocab', 'get_tokens', 'normalize_answer']
+__all__ = ['get_vocab', 'get_tokens', 'normalize_answer', 'mask_padded_tokens']
 
 
 def get_vocab(file):
@@ -50,3 +50,8 @@ def get_tokens(s):
     if not s:
         return []
     return normalize_answer(s).split()
+
+
+def mask_padded_tokens(tokens, pad_id):
+    mask = tokens != pad_id
+    return mask
